+++ ChangeLog.rst
@@ -2,6 +2,14 @@
  ChangeLog
 ===========
 
+0.2.0
+=====
+
+* Now only ``user:email`` scope is required for authentication
+  via github.
+* And ``weblocks-auth/github:render-button`` ``weblocks-auth/github:get-scopes``
+  functions was added to request more scopes if required.
+
 0.1.0
 =====